Updates, code to prevent fights

+++ bot.py
@@ -25,51 +25,66 @@ with open(r'bot.cfg') as file:
 
 bot = telebot.TeleBot(cfgopt["auth"]["token"])
 
-@bot.message_handler(commands=['start', 'help', 'trustme', 'userid', 'hackme','spam'])
+@bot.message_handler(commands=['start', 'help', 'trustme', 'userid', 'hackme', 'spam', 'nofight', 'debug'])
 def commands_handling(message):
     print(message)
     if(message.chat.type == "private"):
-        if (message.text.startswith("/trustme")):
-            print("trustme " + str(message.from_user.id))
-            result = hashlib.sha256((cfgopt["auth"]["trustedseed"] + str(message.from_user.id)).encode('utf-8')).hexdigest()
-            print("trustme request hash to match " + result)
+        if (message.text.startswith("/start")):
+            bot.reply_to(message, "Hey! Unfortunately this bot dont have any help message, yet")
 
-    if(message.chat.type == "private"):
+#        if (message.text.startswith("/trustme")):
+#            print("trustme " + str(message.from_user.id))
+#            result = hashlib.sha256((cfgopt["auth"]["trustedseed"] + str(message.from_user.id)).encode('utf-8')).hexdigest()
+#            print("trustme request hash to match " + result)
+
         if (message.text.startswith("/userid")):
             bot.reply_to(message, "Hey, your userid is:" + str(message.from_user.id))
 
-    if(message.chat.type == "private"):
-       if (message.text.startswith("/hackme") and cfgopt["auth"]["hackme"] == 1):
-        if (not message.from_user.id in cfgopt["users"]["trusted"]):
-          bot.reply_to(message, "Added you to trusted list")
+        if (message.text.startswith("/debug")):
+            print(message)
+
+        if (message.text.startswith("/trustme") and cfgopt["auth"]["trustme"] == 1 and message.from_user.id not in cfgopt["users"]["newmembers"]):
+         if (not message.from_user.id in cfgopt["users"]["trusted"]):
+          bot.reply_to(message, "Hello Master, i added you to trusted list")
           cfgopt["users"]["trusted"].append(message.from_user.id)
           with open('bot.cfg', 'w') as f:
               toml.dump(cfgopt, f)
-        else:
-            bot.reply_to(message, "You are already trusted")
-       else:
-        bot.reply_to(message, "What are you trying to do?")
+         else:
+            bot.reply_to(message, "You are already trusted or bot have bug")
+
     if(message.chat.type == "supergroup" and message.text.startswith("/spam")):
-        
         #if (message.from_user.id in cfgopt["users"]["trusted"] and message.reply_to_message.from_user.id in cfgopt["users"]["trusted"]):
-        if (message.from_user.id in cfgopt["users"]["trusted"]):
-            user_id = message.from_user.id 
-            user_name = message.from_user.first_name 
+        if (message.from_user.id in cfgopt["users"]["trusted"] and message.reply_to_message is not None):
+            user_id = message.reply_to_message.from_user.id 
+            user_name = message.reply_to_message.from_user.first_name 
             mention = "["+user_name+"](tg://user?id="+str(user_id)+")"            
             bot.reply_to(message.reply_to_message, "Hey," + mention + " ,don't spam please!", parse_mode="Markdown")
             bot.delete_message(message.reply_to_message.chat.id, message.reply_to_message.message_id)
+        bot.delete_message(message.chat.id, message.message_id)
+
+    if(message.chat.type == "supergroup" and message.text.startswith("/nofight")):
+        #if (message.from_user.id in cfgopt["users"]["trusted"] and message.reply_to_message.from_user.id in cfgopt["users"]["trusted"]):
+        if (message.from_user.id in cfgopt["users"]["trusted"] and message.reply_to_message is not None):
+            if (time.time() - message.reply_to_message.date < cfgopt["misc"]["fightage"]):
+              user_id = message.reply_to_message.from_user.id 
+              user_name = message.reply_to_message.from_user.first_name 
+              mention = "["+user_name+"](tg://user?id="+str(user_id)+")"            
+              bot.reply_to(message.reply_to_message, "CALM DOWN OR YOU WILL BE SHOT BY ROBOTIC OVERLORD! Some messages are deleted, users temporary muted (but it is all in log), don't fight please!", parse_mode="Markdown")
+              bot.delete_message(message.reply_to_message.chat.id, message.reply_to_message.message_id)
+              bot.restrict_chat_member(message.reply_to_message.chat.id, message.reply_to_message.from_user.id, until_date=time.time()+ 300)
 
         bot.delete_message(message.chat.id, message.message_id)
 
-@bot.message_handler(func=lambda message: message.forward_from_chat, content_types=["text", "photo", "video"])
-def posts_from_channels(message):
-    bot.reply_to(message, 'Hey, don\'t send this to our chat please')
-    bot.delete_message(message.chat.id, message.message_id)
-#    print(message)
 
-@bot.message_handler(func=lambda message: message.forward_from_message_id, content_types=["text", "photo", "video"])
-def posts_from_channels(message):
-    bot.reply_to(message, 'Msg fwd heh?')
+#@bot.message_handler(func=lambda message: message.forward_from_chat, content_types=["text", "photo", "video"])
+#def posts_from_channels(message):
+#    bot.reply_to(message, 'Hey, don\'t send this to our chat please')
+#    bot.delete_message(message.chat.id, message.message_id)
+
+
+#@bot.message_handler(func=lambda message: message.forward_from_message_id, content_types=["text", "photo", "video"])
+#def posts_from_channels(message):
+#    bot.reply_to(message, 'Msg fwd heh?')
 
 
 # TODO: Check new member id to estimate age?
@@ -81,10 +96,13 @@ def posts_from_channels(message):
 def new_chat_member_handling(message):
     bot.reply_to(message, "Welcome to our chat!")
     cfgopt["users"]["newmembers"].append(message.from_user.id)
+    with open('bot.cfg', 'w') as f:
+      toml.dump(cfgopt, f)
+
 
 @bot.message_handler(func=lambda m: True, content_types=["text", "photo", "video"])
 def handle_all(message):
-    if (message.from_user.id in cfgopt["users"]["newmembers"] and len(message.entities) > 0):
+    if (message.from_user.id in cfgopt["users"]["newmembers"] and message.entities != None and len(message.entities) > 0):
       user_id = message.from_user.id 
       user_name = message.from_user.first_name 
       mention = "["+user_name+"](tg://user?id="+str(user_id)+")"            
@@ -100,8 +118,6 @@ def handle_all(message):
 try:
   bot.polling(none_stop=True)
 except Exception as e:
-
         logger.error(e)
-
-        time.sleep(15)
+        time.sleep(5)
